3/requirements.txt
python-docx==0.8.11
pdfplumber==0.10.3
reportlab==4.0.9
nltk==3.8.1
tqdm==4.66.2
PyPDF2==3.17.1
